Refactor chat into hook and subcomponents
